# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Doc API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/products": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "List of products",
                "operationId": "633e179f67b45f73527b111f8460125c",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Some product"
                                                    },
                                                    "price": {
                                                        "type": "integer",
                                                        "example": 100
                                                    },
                                                    "category": {
                                                        "type": "string",
                                                        "example": "some category"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Product"
                ],
                "summary": "Create",
                "operationId": "c0b7bcdc5864ad8bf7876b64de2a23e1",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "name": {
                                                "type": "string",
                                                "example": "Some product"
                                            },
                                            "price": {
                                                "type": "integer",
                                                "example": 100
                                            },
                                            "category_id": {
                                                "type": "integer",
                                                "example": 1
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "name": {
                                                    "type": "string",
                                                    "example": "Some product"
                                                },
                                                "price": {
                                                    "type": "integer",
                                                    "example": 100
                                                },
                                                "category": {
                                                    "type": "string",
                                                    "example": "some category"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/products/{product}": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "Show product",
                "operationId": "7160b436dad66f47ddda6974cb196d52",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "ID product",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "name": {
                                                    "type": "string",
                                                    "example": "Some product"
                                                },
                                                "price": {
                                                    "type": "integer",
                                                    "example": 100
                                                },
                                                "category": {
                                                    "type": "string",
                                                    "example": "some category"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Product"
                ],
                "summary": "Delete product",
                "operationId": "44751ca387e00405e7f0cb4a9291f029",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "ID product",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 204
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "Product"
                ],
                "summary": "Edit product",
                "operationId": "b7ebfed8bdc682613a9164acd32874f1",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "ID product",
                        "required": true,
                        "example": 2
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "name": {
                                                "type": "string",
                                                "example": "Edit some product"
                                            },
                                            "price": {
                                                "type": "integer",
                                                "example": 100
                                            },
                                            "category_id": {
                                                "type": "integer",
                                                "example": 1
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "name": {
                                                    "type": "string",
                                                    "example": "Edit some product"
                                                },
                                                "price": {
                                                    "type": "integer",
                                                    "example": 100
                                                },
                                                "category": {
                                                    "type": "string",
                                                    "example": "some category"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}